cartoonify_image/requirements.txt
1.Numpy
2.OpenCv-python
3.PIL(Python Image Library)
4.tkinter